add live mode as will and static modes
